requirements.txt
streamlit
geopy
copernicusmarine
google-generativeai